converting images | files permission

+++ controllers/tasksController.js
@@ -1,6 +1,6 @@
 const asyncHnadler = require('express-async-handler');
 const taskModel = require('../models/task')
-
+const userModel = require('../models/user') 
 
 const getTasks = asyncHnadler(async (req, res) => {
     const tasks = await taskModel.find()
@@ -25,6 +25,19 @@ const updateTask =  asyncHnadler(async (req, res) => {
         res.status(400)
         throw new Error('ttask not exist')
     }
+    // check for user 
+    const user = await userModel.findById(req.user.id)
+    if (!user) {
+        res.status(400)
+        throw new Error('user not exist')
+    }
+
+    // check if the task owner who want to update
+
+    if (user.id !== task.user.toString()) {
+        res.status(401)
+        throw new Error('user not authorized')   
+    }
     const updatedTask = await taskModel.findByIdAndUpdate(req.params.id, updateTask,{new : true})
     res.status(200).json(updatedTask);
    
@@ -37,7 +50,19 @@ const deleteTask =  asyncHnadler(async (req, res) => {
         res.status(400)
         throw new Error('task not exist')
     }
+    // check for user 
+    const user = await userModel.findById(req.user.id)
+    if (!user) {
+        res.status(400)
+        throw new Error('user not exist')
+    }
 
+    // check if the task owner who want to update
+
+    if (user.id !== task.user.toString()) {
+        res.status(401)
+        throw new Error('user not authorized')   
+    }
     const deletedTask = await taskModel.findByIdAndDelete(req.params.id)
     res.status(200).json(deleteTask)
 })
